+++ rust/src/lib.rs
@@ -16,6 +16,8 @@ mod serde_utils;
 mod types;
 mod utils;
 
+pub use utils::counter;
+
 #[cfg(feature = "blocking")]
 #[cfg_attr(docsrs, doc(cfg(feature = "blocking")))]
 pub mod blocking;